% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{calibration_ellipsoid}
\alias{calibration_ellipsoid}
\alias{calibration_ellipsoid-class}
\title{An S4 class to organize data and results of calibration_ellipsoid objects}
\description{
An S4 class to organize data and results of calibration_ellipsoid objects
}
\section{Slots}{

\describe{
\item{\code{method}}{("character")}

\item{\code{data}}{list of data.frames with the complete set of occurrences, the
training set, and the testing set.}

\item{\code{variable_sets}}{list of two objects, a RasterStack with variables, and a
list of character vectors with sets of variables.}

\item{\code{level}}{("numeric")}

\item{\code{results}}{a data.frame of all results of the model evaluation.}

\item{\code{selection_criteria}}{("character")}

\item{\code{selected_parameters}}{data.frame of selected paramete settings.}
}}

\examples{
showClass("calibration_ellipsoid")
}
