% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_guide.R
\name{calibration_guide}
\alias{calibration_guide}
\title{A guide for ellipsoid ENM calibration}
\usage{
calibration_guide(data, species, longitude, latitude, raster_layers,
  accessible_area, output_directory = "ellipsenm_calibration")
}
\arguments{
\item{data}{data.frame of occurrence records. Columns must be: species,
longitude, and latitude. Optionally, if \code{raster_layers} is not defined,
\code{data} must include more columns containing the values of at least two
variables to be used for fitting ellipsoid* models.}

\item{species}{(character) name of the column with the name of the species.}

\item{longitude}{(character) name of the column with longitude data.}

\item{latitude}{(character) name of the column with latitude data.}

\item{raster_layers}{RasterStack of at least two environmental variables to be
extracted using geographic coordinates present in \code{data}. If not provided
data must include additional columns containing values of variables to fit
ellipsoid* models.}

\item{accessible_area}{SpatialPolygon* or RasterLayer to be used in masking
\code{raster_layers} according to a hypothesis of areas that have been
accessible to the species during a relevant period of time. See details.}

\item{output_directory}{name of the folder were all results will be written.
This helps to organize data and results and avoid confussion.}
}
\value{
A folder containing an Rmarkdown file that serves as a guide to perform the
process of model calibration. The Rmarkdown will be automatically open after
running the function and all processes can be performed there.
}
\description{
calibration_guide generates an Rmarkdown file with information,
instructions, and details, to guide users through the process of model
calibration.
}
\details{
\code{accessible_area}: if the user does not have a hypothesis of this area,
different options to create one exist. Four potential options to define this
area are: (1) a manual definition of accessible areas for the species; (2) the
use of geographic regions potentially relevant (e.g., Ecorregions, Biomes,
basins), generally they require further processing; (3) using other polygons
in the geography (e.g., buffers, convex hulls, or concave hulls); (4) finally,
a definition of the M can be done via simulations of dispersal based on the
species ecological characteristics. For the latter see (****).

See functions \code{\link{buffer_area}}, \code{\link{convex_area}},
\code{\link{concave_area}}, and \code{\link{polygon_selection}}, for options
to create accessible areas based on geographic polygons and selection of
relevant parts of existing SpatialPolygon areas.

For creation of accessible areas from simulations see function M_simulation
from the grinnell package \url{https://github.com/fmachados/grinnell}.
}
