% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cluster_split}
\alias{cluster_split}
\title{Split data based on clusters}
\usage{
cluster_split(
  data,
  cluster_method = "hierarchical",
  split_distance = NULL,
  n_kmeans = NULL
)
}
\arguments{
\item{data}{data.frame of occurrence records containing at least longitude
and latitude columns.}

\item{cluster_method}{(character) name of the method to be used for clustering
the occurrences. Options are "hierarchical" and "k-means"; default =
"hierarchical".}

\item{split_distance}{(numeric) distance in km that will be considered as the
limit of connectivity among polygons created with clusters of occurrences.
This parameter is used when \code{cluster_method} = "hierarchical".
Default = NULL.}

\item{n_kmeans}{(numeric) if \code{split} = TRUE, number of clusters in which
the species occurrences will be grouped when using the "k-means"
\code{cluster_method}. Default = NULL.}
}
\description{
Split data based on clusters
}
\details{
The \code{cluster_method} must be chosen based on the spatial
configuration of the species occurrences. Both methods make distinct assumptions
and one of them may perform better than the other depending on the spatial
pattern of the data.

The k-means method, for example, perfomrs better when the following assumptions
are fulfilled: Clusters are spatially grouped—or “spherical” and Clusters are
of a similar size. Owing to the nature of the hierarchical clustering algorithm
it may take more time than the k-means method. Both methods make assumptions
and they may work well on some data sets, and fail on others.

Another important factor to consider is that the k-means method allways starts
with a random choice of cluster centers, thus it may end in different results
on different runs. That may be problematic when trying to replicate your
methods. With hierarchical clustering, most likely the same clusters can be
obtained if the process is repeated.

For more information on these clustering methods see Aggarwal and Reddy (2014)
\url{https://goo.gl/RQ2ebd}.
}
