% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concave_area.R
\name{concave_area}
\alias{concave_area}
\title{Concave hull polygons based on occurrences}
\usage{
concave_area(data, longitude, latitude, length_threshold = 5,
             split = FALSE, n_kmeans = NULL, buffer_distance = NULL,
             raster_layers = NULL, clip = FALSE, mask = FALSE,
             save = FALSE, name = "calib_area_concave")
}
\arguments{
\item{data}{data.frame of occurrence records containing at least longitude
and latitude columns.}

\item{longitude}{(character) name of the column with longitude data.}

\item{latitude}{(character) name of the column with latitude data.}

\item{length_threshold}{(numeric) distance in km to define when a segment
length stops being considered for further detalization. Higher values result
in simpler shapes. Default = 5.}

\item{split}{(logical) if TRUE, a distance (for hierarchical clustering) or
a number (for K-means clustering) is used to separate distinct chunks of
data Recommended when the species of interest has a disjunct distribution.
Default = FALSE.}

\item{n_kmeans}{(numeric) if \code{split} = TRUE, number of clusters in which
the species occurrences will be grouped using the "k-means" method.
Default = NULL.}

\item{buffer_distance}{(numeric) distance in km to be used to create a buffer
for the concave hull. Default = NULL.}

\item{raster_layers}{optional RasterStack to be used in restricting the
resulting SpatialPolygon and in preparing variables for further processing.
Default = NULL.}

\item{clip}{(logical) whether or not to clip polygons considering boundaries
of layers in \code{raster_layers}. Using this option increases the time of
processing considerably. Default = FALSE.}

\item{mask}{(logical) whether or not to mask the \code{raster_layers} to the
area created with thsi function. Default = FALSE.}

\item{save}{(logical) whether or not to write the results in the working
directory. Default = FALSE.}

\item{name}{(character) name of a folder to be written with the results if
\code{save} = TRUE. Default = "calib_area_concave".}
}
\value{
If raster layers are masked, a lits containing a SpatialPolygonDataFrame and
a RasterStack of the masked layers.

If raster layers are not masked, a SpatialPolygonDataFrame.

If \code{save} = TRUE, results are written in a folder named as in \code{name}.
}
\description{
concave_area helps in creating concave polygons based on
occurrences, buffering polygons, masking raster layers, and writing results
if needed.
}
\examples{
# reading data
occurrences <- read.csv(system.file("extdata", "occurrences.csv",
                                    package = "ellipsenm"))

# producing simple concave polygons
cv_area <- concave_area(data = occurrences, longitude = "longitude",
                       latitude = "latitude")

sp::plot(cv_area)
points(occurrences[, 2:3])

# producing concave polygons with buffers
cv_area1 <- concave_area(data = occurrences, longitude = "longitude",
                        latitude = "latitude", buffer_distance = 50)

sp::plot(cv_area1)
points(occurrences[, 2:3])

# producing concave polygons splitted considering clusters
cv_area2 <- concave_area(data = occurrences, longitude = "longitude",
                        latitude = "latitude", split = TRUE, n_kmeans = 2,
                        buffer_distance = 5)

sp::plot(cv_area2)
points(occurrences[, 2:3])

# producing concave polygons, masking layers
vars <- raster::stack(list.files(system.file("extdata", package = "ellipsenm"),
                                 pattern = "bio", full.names = TRUE))

cv_area3 <- concave_area(data = occurrences, longitude = "longitude",
                        latitude = "latitude", buffer_distance = 50,
                        raster_layers = vars, mask = TRUE)

raster::plot(cv_area3$masked_variables[[1]])
sp::plot(cv_area3$calibration_area, add = TRUE)
points(occurrences[, 2:3])

# producing concave polygons, masking layers, and saving results
cv_area4 <- concave_area(data = occurrences, longitude = "longitude",
                        latitude = "latitude", buffer_distance = 50,
                        raster_layers = vars, mask = TRUE, save = TRUE,
                        name = "concave_area")

# check directory
dir()
}
