% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_area.R
\name{convex_area}
\alias{convex_area}
\title{Convex hull polygons based on occurrences}
\usage{
convex_area(data, longitude, latitude, split = FALSE, n_kmeans = NULL,
  buffer_distance = NULL, raster_layers = NULL, clip = FALSE,
  mask = FALSE, save = FALSE, name = "calib_area_convex")
}
\arguments{
\item{data}{data.frame of occurrence records containing at least longitude
and latitude columns.}

\item{longitude}{(character) name of the column with longitude data.}

\item{latitude}{(character) name of the column with latitude data.}

\item{split}{(logical) if TRUE, a distance (for hierarchical clustering) or
a number (for K-means clustering) is used to separate distinct chunks of
data Recommended when the species of interest has a disjunct distribution.
Default = FALSE.}

\item{n_kmeans}{(numeric) if \code{split} = TRUE, number of clusters in which
the species occurrences will be grouped using the "k-means" method.
Default = NULL.}

\item{buffer_distance}{(numeric) distance in km to be used to create a buffer
for the convex hull. Default = NULL.}

\item{raster_layers}{optional RasterStack to be used in restricting the
resulting SpatialPolygon and in preparing variables for further processing.
Default = NULL.}

\item{clip}{(logical) whether or not to clip polygons considering boundaries
of layers in \code{raster_layers}. Using this option increases the time of
processing considerably. Default = FALSE.}

\item{mask}{(logical) whether or not to mask the \code{raster_layers} to the
area created with thsi function. Default = FALSE.}

\item{save}{(logical) whether or not to write the results in the working
directory. Default = FALSE.}

\item{name}{(character) name of a folder to be written with the results if
\code{save} = TRUE. Default = "calib_area_convex".}
}
\value{
If raster layers are masked, a lits containing a SpatialPolygonDataFrame and
a RasterStack of the masked layers.

If raster layers are not masked, a SpatialPolygonDataFrame.

If \code{save} = TRUE, results are written in a folder named as in \code{name}.
}
\description{
convex_area helps in creating convex polygons based on
occurrences, buffering polygons, masking raster layers, and writing results
if needed.
}
\examples{
# data
occurrences <- read.csv(system.file("extdata", "occurrences_comp.csv",
                                    package = "ellipsenm"))

# producing simple convex polygons
cx_area <- convex_area(data = occs, longitude = "LONGITUDE",
                       latitude = "LATITUDE")

# producing convex polygons with buffers
cx_area1 <- convex_area(data = occs, longitude = "LONGITUDE",
                        latitude = "LATITUDE", buffer_distance = 50)

# producing convex polygons splitted considering clusters
cx_area2 <- convex_area(data = occs, longitude = "LONGITUDE",
                        latitude = "LATITUDE", split = TRUE, n_kmeans = 3,
                        buffer_distance = 50)

# producing convex polygons, masking layers
vars <- raster::stack(list.files(system.file("extdata", package = "ellipsenm"),
                                 pattern = "m_bio", full.names = TRUE))

cx_area3 <- convex_area(data = occs, longitude = "LONGITUDE",
                        latitude = "LATITUDE", buffer_distance = 100,
                        raster_layers = vars, mask = TRUE)

# producing convex polygons, masking layers, and saving results
cx_area4 <- convex_area(data = occurrences, longitude = "longitude",
                       latitude = "latitude", buffer_distance = 100,
                       raster_layers = vars, mask = TRUE, save = TRUE,
                       name = "buff_area")
}
