% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{data_overlap}
\alias{data_overlap}
\alias{data_overlap-class}
\title{An S4 class to organize data for data_overlap objects}
\description{
An S4 class to organize data for data_overlap objects
}
\section{Slots}{

\describe{
\item{\code{data}}{data.frame}

\item{\code{main_columns}}{("character")}

\item{\code{method}}{("character")}

\item{\code{level}}{("numeric")}

\item{\code{results}}{a data.frame of all results of the model evaluation.}

\item{\code{variables}}{object of class RasterStack.}
}}

\examples{
showClass("data_overlap")
}
