% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsenm.R
\docType{package}
\name{ellipsenm}
\alias{ellipsenm}
\alias{ellipsenm-package}
\title{ellipsenm: An R package for ecological niche's characterization using ellipsoids}
\description{
ellipsenm implements multiple tools to help users in using ellipsoids to model
ecological niches of species. Handly options for calibrating and selecting
models, producing models with replicates and projections, and assessing niche
overlap are included as part of this package. Other functions implemented here
are useful to perform a series of pre- and post-modeling analyses.
}
\section{Main functions in ellipsenm}{

\code{\link{thin_data}}, \code{\link{split_ata}}, \code{\link{data_subsample}},
\code{\link{buffer_area}}, \code{\link{convex_area}}, \code{\link{concave_area}},
\code{\link{polygon_selection}}, \code{\link{explore_espace}},
\code{\link{ellipsoid_fit}}, \code{\link{ellipsoid_calibration}},
\code{\link{predict}}, \code{\link{ellipsoid_model}},
\code{\link{model_projection}}, \code{\link{partial_roc}},
\code{\link{omission_rate}}
}

