% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid_calibration.R
\name{ellipsoid_calibration}
\alias{ellipsoid_calibration}
\title{Calibration of ellipsoid-based ecological niche models}
\usage{
ellipsoid_calibration(data, species, longitude, latitude, variables,
  methods, level = 95, parallel = FALSE, overwrite = FALSE,
  output_directory = "calibration_results")
}
\arguments{
\item{data}{(character or list) if character, vector of names of csv files
containing all, training, and testing occurrences located in the working
directory; if list, object resulted from \code{\link{split_data}}. Columns of
tables must include: species, longitude, and latitude.}

\item{species}{(character) name of the column with the name of the species.}

\item{longitude}{(character) name of the column with longitude data.}

\item{latitude}{(character) name of the column with latitude data.}

\item{variables}{(character or list) if character, name of a folder containing
subfolders of at least one set of variables; if list, object derived from
\code{\link{prepare sets}}. Sets of variables must contain at least two layers.}

\item{methods}{(character) methods to construct the ellipsoid ecological niche
models to be tested. Available methods are: "covmat", "mve1", and "mve2".
See details of \code{\link{ellipsoid_fit}}.}

\item{level}{(numeric) the confidence level of a pairwise confidence region
for the ellipsoid, expresed as percentage. Default = 95.}
}
\description{
ellipsoid_calibration helps in creating and evaluating multiple
candidate ellipsoid envelop models to find parameter settings that produce
the best results.
}
