% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid_model.R
\name{ellipsoid_model}
\alias{ellipsoid_model}
\title{Fit ellipsoids based on distinct methods}
\usage{
ellipsoid_model(data, species, longitude, latitude, raster_layers,
  method = "mve1", level = 95, replicates = 1,
  replicate_type = "bootstrap", bootstrap_percentage = 50,
  projection_layers = NULL, prediction = "suitability",
  tolerance = 1e-60, output_directory = "ellipsenm_model")
}
\arguments{
\item{data}{data.frame of occurrence records. Columns must be: species,
longitude, and latitude.}

\item{species}{(character) name of the column with the name of the species.}

\item{longitude}{(character) name of the column with longitude data.}

\item{latitude}{(character) name of the column with latitude data.}

\item{raster_layers}{RasterStack of environmental variables to be extracted
using geographic coordinates present in \code{data}.}

\item{method}{(character) method to construct the ellipsoid that characterizes
the species ecological niche. Available methods are: "covmat", "mve1", and
"mve2". See details of \code{\link{ellipsoid_fit}}. Default = "mve1".}

\item{level}{(numeric) the confidence level of a pairwise confidence region
for the ellipsoid, expresed as percentage. Default = 95.}

\item{replicates}{(numeric) number of replicates to perform. Default = 1
produces a single model using all the data.}

\item{replicate_type}{(character) type of replicates to perform. Options are:
"bootstrap" and "jackknife"; default = "bootstrap". See details. Ignored if
\code{replicates} = 1.}

\item{bootstrap_percentage}{(numeric) percentage of data to be bootstrapped
for each replicate. Default = 50. Valid if \code{replicates} > 1 and
\code{replicate_type} = "bootstrap".}

\item{projection_layers}{optional: (character) name of folder containing other
folders with raster layers that represent the scenarios for projections; a
RasterStack of layers respresenting an only scenario for projection; or a
named list of RasterStacks representing multiple scenarios for projection.
See details. Default = NULL.}

\item{prediction}{(character) type of prediction to be made, options are:
"suitability", "mahalanobis", and "both". Default = "suitability".}

\item{tolerance}{the tolerance for detecting linear dependencies.
Default = 1e-60.}

\item{output_directory}{name of the folder were all results will be written.
This avoids saturation of the RAM.}
}
\value{

}
\description{
ellipsoid_fit helps in finding the centroid and matrix that
define an ellipsoid. It uses distinct methods with asumptions that differ
from each other.
}
\details{
\code{replicate_type}

\code{projection_layers}
}
\examples{
occurrences <- read.csv(system.file("extdata", "occurrences_comp.csv",
                                    package = "ellipsenm"))

vars <- raster::stack(list.files(system.file("extdata", package = "ellipsenm"),
                                 pattern = "m_bio", full.names = TRUE))
}
