% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_helpers.R
\name{ellipsoid_volume}
\alias{ellipsoid_volume}
\title{Helper function to calculate niche volume}
\usage{
ellipsoid_volume(n_dimensions, semi_axes_length)
}
\arguments{
\item{n_dimensions}{(numeric) number of dimensions to be considered.}

\item{semi_axes_length}{(numeric) length of ellipsoid axes.}
}
\description{
Helper function to calculate niche volume
}
