% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_espace.R
\name{explore_espace}
\alias{explore_espace}
\title{Exploration of data in enviornmental space}
\usage{
explore_espace(data, species, longitude, latitude, raster_layers,
  name = "e_space_exploration.pdf", open = TRUE)
}
\arguments{
\item{data}{data.frame of occurrence records. Columns must be: species,
longitude, and latitude.}

\item{species}{(character) name of the column with the name of the species.}

\item{longitude}{(character) name of the column with longitude data.}

\item{latitude}{(character) name of the column with latitude data.}

\item{raster_layers}{RasterStack of at least two environmental variables to be
extracted using geographic coordinates present in \code{data}.}

\item{name}{(character) name (including the extention .pdf) of the PDF file to
be created with the exploratory figures. Default = "e_space_exploration.pdf".}

\item{open}{(logical) whether or not to open the file created directly.
Default = TRUE}
}
\value{
A PDF document with indications and figures to explore the data in environmental
space and make desicions on which methods and variables are worth to explore
during model calibration.
}
\description{
explore_espace helps in eploring the patterns of arrangement of
species occurrence data and accessible areas in environmental space.
}
