% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_attribute}
\alias{get_attribute}
\title{Helper funtion to get attributes from ellipsoid lists}
\usage{
get_attribute(ellipsoids, attribute = "method")
}
\arguments{
\item{ellipsoids}{list of ellipsoid* objects.}

\item{attribute}{(character) name of the attribute to be obtained from elements
in \code{ellipsoids}. Options are: method, centroid, covariance_matrix, level,
niche_volume, semi_axes_length, and axes_coordinates. Default = "method".}
}
\description{
Helper funtion to get attributes from ellipsoid lists
}
