% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide.R
\name{guide}
\alias{guide}
\title{Creation of HTML reports from ellipsenm main processes}
\usage{
guide(guide_type, output_directory = getwd())
}
\arguments{
\item{guide_type}{(character) type of report to be produced. Options are:
"calibration", "enm", "whole_process", "overlap".}

\item{output_directory}{(character) name of the directory where the HTML file
will be produced. Default = working directory.}
}
\value{
An Rmarkdown file that serves as a guide to perform any of the three processes
defined in \code{guide_type}. The Rmarkdown will be automatically open after
running the function and all processes can be performed there.
}
\description{
report creates an HTML file that summarizes all outputs from
the main processes that the ellipsenm package performs.
}
