% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbased_mve.R
\name{mbased_mve}
\alias{mbased_mve}
\title{Fit moment-based minimum volumen ellipsoids}
\usage{
mbased_mve(data, fitting_tolerance = 0.001)
}
\arguments{
\item{data}{matrix of values to be used for fitting the ellipsoid. Columns
represent dimensions and rows observations.}

\item{fitting_tolerance}{(numeric) proportion of error allowed when checking
if the ellipsoid incloses all values considered to fit it. Default = 0.001.}
}
\value{
A named list containing the values for the centroid of th ellipsoid and a
matrix of covariance.
}
\description{
mbased_mve helps in finding the centroid and matrix that
define a minimum volume ellipsoid as proposed by Moshtagh (2005).
}
\details{
The algorithm used to fit the ellipsoids here was implemented by
Moshtagh (2005) and based on initial work of Khachiyan (1996) and
Rocha et al. (2002).

Details about to the algortihm can be found in \url{https://bit.ly/2XYWlVT},
\url{https://bit.ly/2x6aR2s}, and \url{https://bit.ly/2KrxE1g}.
}
\examples{
# reading data
occurrences <- read.csv(system.file("extdata", "occurrences.csv",
                                    package = "ellipsenm"))

# raster layers of environmental data
vars <- raster::stack(list.files(system.file("extdata", package = "ellipsenm"),
                                 pattern = "bio", full.names = TRUE))

data <- raster::extract(vars, occurrences[, -1])

mvellipsoid <- mbased_mve(data)
}
