% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_ellipsoid.R
\name{mmm_ellipsoid}
\alias{mmm_ellipsoid}
\title{Mean, min, and max ellipsoids}
\usage{
mmm_ellipsoid(ellipsoid_list)
}
\arguments{
\item{ellipsoid_list}{list of ellipsoid objects created with the function
\code{\link{ellipsoid_fit}}.}
}
\description{
mmm_ellipsoid helps in finding the mean, minimum, and maximum
ellipsoids of a list of ellipsoid objects.
}
\details{
Minimum and maximum ellipsoids are calculated based on the volumes of all
elements in \code{ellipsoid_list}.

The mean ellipsoid is calculated by finsing the mean of all centroids and all
covariance matrices of all elements in \code{ellipsoid_list}.
}
