% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid_fit.R
\name{ndata_quantile}
\alias{ndata_quantile}
\title{Helper function to calculate quantiles}
\usage{
ndata_quantile(n_data, level)
}
\arguments{
\item{n_data}{(numeric) total number of data to be considered.}

\item{level}{(numeric) percentage of data to be considered when creating the
ellipsoid that characterizes the species ecological niche. Default = 95.}
}
\description{
Helper function to calculate quantiles
}
