% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{occ_blocksplit}
\alias{occ_blocksplit}
\title{Split occurrences in training and testing data using blocks}
\usage{
occ_blocksplit(data, longitude, latitude, train_proportion = 0.75)
}
\arguments{
\item{data}{matrix or data.frame with the occurrences to be split. Columns
may vary but species, longitude, and latitue are recommended.}

\item{longitude}{(character) name of the column with longitude data.}

\item{latitude}{(character) name of the column with latitude data.}

\item{train_proportion}{(numeric) proportion of data to be used as training
occurrences. Available options are 0.25, 0.5, and 0.75. The remaining data
will be used for testing. Default = 0.75.}
}
\value{
List with all occurrences (all), training occurrences (train), and testing
(test) occurrences.
}
\description{
occ_blocksplit splits a set of occurrences to obtain training and
testing data randomly.
}
