% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{occ_randsplit}
\alias{occ_randsplit}
\title{Split occurrences randomly in training and testing data}
\usage{
occ_randsplit(data, train_proportion = 0.5)
}
\arguments{
\item{data}{matrix or data.frame with the occurrences to be split. Columns
may vary but species, longitude, and latitue are recommended.}

\item{train_proportion}{(numeric) proportion (from 0 to 1) of data to be used
as training occurrences. The remaining data will be used for testing.
Default = 0.5.}
}
\value{
List with all occurrences (all), training occurrences (train), and testing
(test) occurrences.
}
\description{
occ_randsplit splits a set of occurrences to obtain training and
testing data randomly.
}
