% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{overlap_metrics}
\alias{overlap_metrics}
\title{Helper function to compute overlap metrics}
\usage{
overlap_metrics(
  comparison_matrix,
  background,
  mahalanobis,
  suitability,
  return_background = TRUE
)
}
\arguments{
\item{comparison_matrix}{matrix resulted from using the \code{\link[utils]{combn}}
function to obtain the pairwise overlap of all niches to be compared.}

\item{background}{matrix of environmental variables to be used as background.}

\item{mahalanobis}{matrix of mahalanobis distances calculated for all niches
to be compared.}

\item{suitability}{matrix of suitability values calculated for all niches to
be compared.}
}
\value{
A list of two data.frames containing metrics of overlap and other data useful
for plotting results.
}
\description{
Helper function to compute overlap metrics
}
