% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{overlap_object}
\alias{overlap_object}
\title{Helper funtion to create data_overlap objects}
\usage{
overlap_object(data, species, longitude, latitude, method = "covmat",
  level = 95, variables = NULL)
}
\arguments{
\item{data}{data.frame of species' occurrence records. Columns must include
species, longitude, and latitude.  Optionally, if \code{variables} is a matrix
or data.frame, \code{data} must include more columns containing the values of
at least two variables to be used for fitting ellipsoid* models.}

\item{method}{(character) method to construct the ellipsoid that characterizes
the species ecological niche. Available methods are: "covmat", "mve1", and
"mve2". See details. Default = "covmat".}

\item{level}{(numeric) the confidence level of a pairwise confidence region
for the ellipsoid, expresed as percentage. Default = 95.}

\item{variables}{(optional) RasterStack, matrix, or data.frame of at least two
variables to represent a set of conditions relevant for overlap analyses.}
}
\value{
A data_overlap object for posterior use in overlap analyses.
}
\description{
Helper funtion to create data_overlap objects
}
