% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_random.R
\name{overlap_random}
\alias{overlap_random}
\title{Overlap analyses for random sampled data}
\usage{
overlap_random(background, sample_size, method, level, overlap_type = "all",
               replicates = 1000)
}
\arguments{
\item{background}{list containing the backgrounds to be used. If \code{overlap_type}
= "all", background for "full" overlap must be first, and background for "union"
overlap, second.}

\item{sample_size}{(numeric) vector of sample sizes for the species under
analyses that should be sampled in each replicate.}

\item{method}{(character) vector of methods to construct the ellipsoids that
characterize the species ecological niches.}

\item{level}{(numeric) vector of the confidence levels of a pairwise
confidence region for the ellipsoids, expresed as percentage.}

\item{overlap_type}{(character) type of overlap to be measured. Options are:
"all", "full", and "back_union". Default = "all". See details.}

\item{replicates}{(numeric) number of replicates to be performed during the
significance test; default = 1000.}
}
\value{
A list of data.frames with the overlap metrics calculated for each replicate,
according to the \code{overlap_type} selected.
}
\description{
overlap_random performs analyses of overlap on ellipsoidal objects
derived from random samples of their background.
}
\details{
Details about the \code{method} can be revised in \code{\link{ellipsoid_fit}}.

Details about the overlap analyses can be checked in \code{\link{ellipsoid_overlap}}.
}
