% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overlap_sig.R
\name{plot_overlap_sig}
\alias{plot_overlap_sig}
\title{Plot results of significance tests for niche overlap analyses}
\usage{
plot_overlap_sig(object, niches = c(1, 2), main = "",
                 xlab = "Overlap", ylab = "Frequency",
                 breaks = 30, bar_col = "gray65",
                 cl_col = "darkgreen", observed_col = "blue",
                 cl_lty = 2, observed_lty = 1, cl_lwd = 2,
                 observed_lwd = 2, xlim = NULL)
}
\arguments{
\item{object}{overlap_ellipsoid object resulted from using the function
\code{\link{ellipsoid_overlap}}.}

\item{niches}{(numeric) pair of integer numbers denoting the niches for
which results will be plotted. Default = c(1, 2).}

\item{main}{(character) plot main title, default = "".}

\item{xlab}{(character) label of x axis, default = "Overlap".}

\item{ylab}{(character) label of y axis, default = "Frequency".}

\item{breaks}{type of breaks as in \code{\link[graphics]{hist}}, default = 30.}

\item{bar_col}{color of histogram bars, default = "gray65".}

\item{cl_col}{color of confidence limit line, default = "darkgreen".}

\item{observed_col}{color of line representing observed value of overlap,
default = "blue".}

\item{cl_lty}{type of line for confidence limit, default = 2.}

\item{observed_lty}{type of line for observed value of overlap, default = 1.}

\item{cl_lwd}{thickness of line for confidence limit, default = 2.}

\item{observed_lwd}{thickness of line for observed value of overlap,
default = 2.}

\item{xlim}{limits of x axis, the default, NULL, derives limits from data.}
}
\value{
A plot with results from tests of statistical significance.
}
\description{
plot_overlap_sig helps to create a plot to interpret results
from statistical significance tests in overlap analyses.
}
