% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_rep.R
\docType{methods}
\name{predict,ellipsoid_model_rep-method}
\alias{predict,ellipsoid_model_rep-method}
\title{Predict suitability derived from replicated ellipsoid envelope models}
\usage{
\S4method{predict}{ellipsoid_model_rep}(object, projection_layers,
  prediction = "suitability", return_numeric = FALSE,
  tolerance = 1e-60, name = NULL, format, overwrite = FALSE)
}
\arguments{
\item{object}{a fitted object of class ellipsoid_model_rep.}

\item{projection_layers}{RasterStack or matrix of variables representing
environmental conditions of the scenario to which \code{object} will be
projected. See details.}

\item{prediction}{(character) type of prediction to be made, options are:
"suitability", "mahalanobis", and "both". Default = "suitability".}

\item{return_numeric}{(logical) whether or not to return values of mahalanobis
distance and suitability as part of the results (it depends on the type of
\code{prediction} selected). Default = FALSE.}

\item{tolerance}{the tolerance for detecting linear dependencies.
Default = 1e-60.}

\item{name}{(character) optional, name of the file to be writen. Must include
format extension (e.g., ".tif"). When defined, raster predictions are not
returned to the environment. Default = NULL. See detals.}

\item{format}{(charater) if \code{name} is defined, file type to be written.
Must correspond with format extension in \code{name}.
See \code{\link[raster]{writeFormats}}.}

\item{overwrite}{(logical) if \code{name} is defined, whether or not to
overwrite an exitent file with the exact same name. Default = FALSE.}
}
\value{
An ellipsoid_model_rep with new predictions.
}
\description{
predicts suitability values based on mahalanobis distances
based on a centroid and a covariance matrix.
}
\details{
Predictions for all replicates will be performed. If \code{name} is defined,
a prefix starting in 1 will be added to each replicate. After replicate number
the word "suitability" or "mahalanobis" will be added depending on the type
of prediction defined in \code{prediction}.

For \code{projection_layers} variables can be given either as a RasterStack
or as a matrix. If a matrix is given each column represents a variable and
predictions are returned only as numeric vectors. In both cases, variable
names must match exactly the order and name of variables used to create
\code{object}.
}
\examples{
# reading data
occurrences <- read.csv(system.file("extdata", "occurrences.csv",
                                    package = "ellipsenm"))

# raster layers of environmental data
vars <- raster::stack(list.files(system.file("extdata", package = "ellipsenm"),
                                 pattern = "bio", full.names = TRUE))

# getting values of variables in data
occurrences1 <- cbind(occurrences[, 2:3], raster::extract(vars, occurrences[, 2:3]))

# subsampling data for construction of multiple ellipsoids
subsamples <- data_subsample(data = occurrences1, replicates = 10,
                             replicate_type = "bootstrap")

# fitting ellipsoids for the 10 subsamples
ellipsoids <- lapply(subsamples, function (x) {
  ellipsoid_fit(data = x, longitude = "longitude",
                latitude = "latitude", method = "mve1",
                level = 99)
})
length(ellipsoids)
names(ellipsoids) <- paste0("replicate_", 1:10)

# creating an ellipsoid_model_rep object with replicates and mean, min, max
ell_rep <- new("ellipsoid_model_rep",
               ellipsoids = ellipsoids) # again some slots empty here

# predicting suitability
prediction_rep <- predict(object = ell_rep, projection_layers = vars,
                          prediction = "suitability")

# predicting mahalanobis distance
prediction_rep <- predict(object = ell_rep, projection_layers = vars,
                          prediction = "mahalanobis")

# predicting both things
prediction_rep <- predict(object = ell_rep, projection_layers = vars,
                          prediction = "both")
}
