% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sets.R
\name{prepare_sets}
\alias{prepare_sets}
\title{Prepare sets of variables for calibration}
\usage{
prepare_sets(variables, sets = NULL, all_combinations = FALSE,
             minvar_perset = 2, format_in = NULL, save = FALSE,
             format_out = "GTiff", overwrite = FALSE,
             output_directory = "Variable_sets")
}
\arguments{
\item{variables}{(character or RasterStack) if character, name of the folder
containing only the variables to be subsetted; if RasterStack, stack of
variables to be subsetted.}

\item{sets}{named list of character vectors with the names of the variables
per each subset. Default = NULL.}

\item{all_combinations}{(logical) whether or not to prepare sets based on all
potential combinations of two or more \code{variables}. Ignored if sets is
defined and default = FALSE changes to TRUE if sets = NULL.}

\item{minvar_perset}{(integer) if \code{all_combinations} = TRUE and sets is
not defined, minimum number of variables per combination. This number must
be > 1. Default = 2.}

\item{format_in}{(character) if \code{variables} is character, format of the
variables found in the folder. Default = NULL.}

\item{save}{(logical) wheter or not to write sets of variables in a subfolder
in the working directory. Default = FALSE.}

\item{format_out}{(character) format of the variables to be written in
\code{output_directory} if \code{save} = TRUE. Default = "GTiff".}

\item{overwrite}{(logical) whether or not to overwrite exitent results in
\code{output_directory} if \code{save} = TRUE. Default = FALSE.}

\item{output_directory}{name of the folder were subsets will be written if
\code{save} = TRUE. Default = "Variable_sets".}
}
\value{
A list of character vectors containing the names of the variables to be used
in each set. A folder with subfolders (sets) and variables subsetted.
}
\description{
Prepare sets of variables for calibration
}
