% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_area.R
\name{raster_poly}
\alias{raster_poly}
\title{Helper function to obtain polygon from RasterLayer}
\usage{
raster_poly(raster_layer)
}
\arguments{
\item{raster_layer}{RasterLayer of a region of interest.}
}
\description{
Helper function to obtain polygon from RasterLayer
}
