% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Creation of HTML reports from ellipsenm main processes}
\usage{
report(report_type, prediction, projected, output_directory = getwd())
}
\arguments{
\item{prediction}{(character) if report_type = "enm". Type of prediction that
was performed, oprions are: "suitability", "mahalanobis", and "both".}

\item{projected}{(logical) if report_type = "enm", whether or not the model
was projected to distinct scenarios.}

\item{output_directory}{(character) name of the directory where the HTML file
will be produced. Default = working directory.}

\item{repor_type}{(character) type of report to be produced. Options are:
"calibration", "enm", and "overlap".}
}
\value{
An HTML file summarizing results from the main processes that the ellipsenm
package performs. Some 3D plots from the rgl package may appear.
}
\description{
report creates an HTML file that summarizes all outputs from
the main processes that the ellipsenm package performs.
}
\details{
This function is used along with the \code{\link{ellipsoid_model}},
ellipsoid_calibration, and ellipsoid_overlap functions. To obtain nice reports
a css file with format instructions should exist. See \code{\link{report_format}}.
}
