% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_format.R
\name{report_format}
\alias{report_format}
\title{Helper function to format function reports}
\usage{
report_format(col = "#0000FF", name = "eenm_report_format")
}
\arguments{
\item{col}{(character) hexadecimal code for the main color to be used in the
html report. Default = "#0000FF"; blue.}

\item{name}{(character) name of the CSS file to be written. File format does
not need to be added. Default = "eenm_report_format".}
}
\value{
CSS file containing HTML formatting instructions.
}
\description{
Helper function to format function reports
}
\examples{
# example with green color
report_format(col = "#15AB0A")
}
