% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_area.R
\name{save_areas}
\alias{save_areas}
\title{Helper function to save results from calibration area creation}
\usage{
save_areas(name, area_polygon, area_type, raster_layers = NULL)
}
\arguments{
\item{name}{(character) name of a folder to be written with the results if
\code{save} = TRUE.}

\item{area_polygon}{SpatialPolygonDataFrame to be written in \code{name}.}

\item{area_type}{(character) type of polygona (area) to be written.}

\item{raster_layers}{optional RasterStack of layers to be written in \code{name}.}
}
\description{
Helper function to save results from calibration area creation
}
