% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{select_best}
\alias{select_best}
\title{Helper funtion to select best parameter settings}
\usage{
select_best(calibration_table, selection_criteria = "S_OR_P",
            level = 95, error = 5)
}
\arguments{
\item{calibration_table}{data.frame of results from model calibration in
ellipsenm.}

\item{selection_criteria}{(character) set of criteria to select best models,
options are: "S_OR" (statistical significance and low omission) and
"S_OR_P" (statistical significance, low omission, and low prevalence).
See details. Default = "Sig_OR_Prev".}

\item{level}{(numeric) the confidence level of a pairwise confidence region
for the ellipsoid, expresed as percentage. Default = 95.}

\item{error}{(numeric) value from 0 to 100 to represent the percentage of
potential error (E) that the data could have due to any source of uncertainty.
Default = 5.}
}
\value{
data.frame with the selected parameter settings according to the argument
\code{selection_criteria}.
}
\description{
Helper funtion to select best parameter settings
}
