% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdata_overlap.R
\name{spdata_overlap}
\alias{spdata_overlap}
\title{Prepare species and environmental data for overlap analysis}
\usage{
spdata_overlap(method, data, raster_layers)
}
\arguments{
\item{method}{(character) method to contruct the ellipsoid that characterizes
the species ecological niche. Available methods are: "covmat" and "mve".
See details. Default = "covmat".}

\item{data}{data.frame of occurrence records. Columns must be species,
longitude, and latitude.}

\item{raster_layers}{a RasterStack of environmental data to be used in niche
overlap ananlyses.}
}
\description{
spdata_overlap prepares species and environmental data to perform
niche overlap analyses.
}
\details{
Methods details are as follows:

"covmat"

"mve"
}
\examples{
# data
met <- "covmat"
occurrences <- read.csv(system.file("extdata", "occurrences_thin.csv",
                                    package = "ellipsenm"))
layers <- raster::raster(system.file("extdata", "m_bio1.tif",
                                     package = "ellipsenm"))

sp_over <- spdata_overlap(method = met, data = occurrences,
                          raster_layers = layers)
}
