% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Split occurrence data into training and testing data}
\usage{
split_data(data, train_proportion = 0.5, method = "random",
  save = FALSE, name = "occurrences")
}
\arguments{
\item{data}{data.frame of occurrence records containing at least longitude
and latitude columns.}

\item{train_proportion}{(numeric) proportion (from 0 to 1) of data to be used
as training occurrences. The remaining data will be used for testing.
Default = 0.5.}

\item{method}{(character) method for selecting training and testing data.
Options are: "random" and "block"; default = "random".}

\item{save}{(logical) whether or not to save the results in the working
directory. Default = FALSE.}

\item{name}{(character) if \code{save} = TRUE, name of the csv files to be
written (comon name for all files). A suffix will be added depending on the
type of data: complete set, training set, or testing set of occurrences.
Format (.csv) is automatically added; default = "occurrences".}
}
\value{
A list containing all, training, and testing occurrences. If \code{save} =
TRUE, three csv files will be written in the working directory according to
the name defined in \code{name} plus the suffix _all for all records, _train
for the training set, and _test for the testing set.
}
\description{
split_data splits occurrences into training and testing data
based on distinct methods.
}
\examples{
occurrences <- read.csv(system.file("extdata", "occurrences_thin.csv",
                                    package = "kuenm"))

# random split 50\% for trainig and 50\% for testing
data_split <- split_data(data = occs, train.proportion = 0.5)
}
