% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{summary,calibration_ellipsoid-method}
\alias{summary,calibration_ellipsoid-method}
\title{Summary of attributes and results}
\usage{
\S4method{summary}{calibration_ellipsoid}(object)
}
\arguments{
\item{object}{calibration_ellipsoid object.}
}
\value{
A written summary.
}
\description{
Summary of attributes and results
}
