% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{summary,data_overlap-method}
\alias{summary,data_overlap-method}
\title{Summary of attributes and results}
\usage{
\S4method{summary}{data_overlap}(object)
}
\arguments{
\item{object}{data_overlap object.}
}
\value{
A written summary.
}
\description{
Summary of attributes and results
}
