% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{summary,ellipsoid-method}
\alias{summary,ellipsoid-method}
\title{Summary of attributes and results}
\usage{
\S4method{summary}{ellipsoid}(object)
}
\arguments{
\item{object}{ellipsoid or ellipsoid_model_sim object.}
}
\value{
A written summary.
}
\description{
Summary of attributes and results
}
