% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{summary,overlap_ellipsoid-method}
\alias{summary,overlap_ellipsoid-method}
\title{Summary of attributes and results}
\usage{
\S4method{summary}{overlap_ellipsoid}(object)
}
\arguments{
\item{object}{overlap_ellipsoid object.}
}
\value{
A written summary.
}
\description{
Summary of attributes and results
}
