% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,ellipsoid-method}
\alias{summary,ellipsoid_model_rep-method}
\alias{summary,calibration_ellipsoid-method}
\alias{summary,data_overlap-method}
\alias{summary,overlap_ellipsoid-method}
\title{Summary of attributes and results}
\usage{
\S4method{summary}{ellipsoid}(object)

\S4method{summary}{ellipsoid_model_rep}(object)

\S4method{summary}{calibration_ellipsoid}(object)

\S4method{summary}{data_overlap}(object)

\S4method{summary}{overlap_ellipsoid}(object)
}
\arguments{
\item{object}{object of class ellipsoid*, ellipsoid_model_rep,
calibration_ellipsoid, data_overlap, or overlap_ellipsoid.}
}
\value{
A written summary.
}
\description{
Summary of attributes and results
}
