% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_multicollinearity.R
\name{vif_groups}
\alias{vif_groups}
\title{Helper function to find non-collinear groups of variables based on VIF}
\usage{
vif_groups(data, longitude, latitude, raster_layers,
  raster_values = NULL)
}
\description{
Helper function to find non-collinear groups of variables based on VIF
}
