% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin_data.R
\name{wgs_aeqd}
\alias{wgs_aeqd}
\title{Helper function to project points}
\usage{
wgs_aeqd(data, longitude, latitude)
}
\arguments{
\item{data}{data.frame of occurrence records containing at least longitude and
latitude columns.}

\item{longitude}{(character) name of the column with longitude data.}

\item{latitude}{(character) name of the column with latitude data.}
}
\value{
A SpatialPointsDataFrame object projected to azimuthal equidistant projection.
}
\description{
Helper function to project points
}
