% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_subsample.R
\name{data_subsample}
\alias{data_subsample}
\title{Helper function to subsample occurrence data}
\usage{
data_subsample(data, replicates = 1, replicate_type = "bootstrap",
               bootstrap_percentage = 50)
}
\arguments{
\item{data}{data.frame of occurrence records. Columns must be longitude and
latitude.}

\item{replicates}{(numeric) number of subsamples to be produced. If
\code{replicate_type} = "jackknife" the maximum number of subsamples produced
equal the number of records in \code{data}. Default = 1 returns the complete
set of data as the only element of a list.}

\item{replicate_type}{(character) name of the sampling method to be used.
Options are "bootstrap" or "jackknife". Default = "bootstrap".}

\item{bootstrap_percentage}{(numeric) percentage of data to be bootstrapped
for each replicate. Default = 50. Valid if \code{replicates} > 1 and
\code{replicate_type} = "bootstrap".}
}
\description{
Helper function to subsample occurrence data
}
\examples{
# reading data
occurrences <- read.csv(system.file("extdata", "occurrences.csv",
                                    package = "ellipsenm"))

# subsampling by bootstrap
subsamples <- data_subsample(data = occurrences[, -1], replicates = 5,
                             replicate_type = "bootstrap", bootstrap_percentage = 70)

lapply(subsamples, head)
lapply(subsamples, dim)

# subsampling by "jackknife" one record out at the time
subsamples1 <- data_subsample(data = occurrences[, -1], replicates = 5,
                              replicate_type = "jackknife")

lapply(subsamples1, head)
lapply(subsamples1, dim)
}
