% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid_fit.R
\name{ellipsoid_fit}
\alias{ellipsoid_fit}
\title{Fit ellipsoids based on distinct methods}
\usage{
ellipsoid_fit(data, longitude, latitude, method = "mve", level,
  raster_layers = NULL)
}
\arguments{
\item{data}{data.frame of occurrence records. Columns must be species,
longitude, and latitude. Other columns are values of environmental variables
to be used as dimensions of the ellipsoid.}

\item{longitude}{(character) name of the column with longitude data.}

\item{latitude}{(character) name of the column with latitude data.}

\item{method}{(character) method to construct the ellipsoid that characterizes
the species ecological niche. Available methods are: "covmat" and "mve".
See details. Default = "mve".}

\item{level}{(numeric) percentage of data to be considered when creating the
ellipsoid that characterizes the species ecological niche.}

\item{raster_layers}{optional RasterStack of environmental variables to be
extracted using geographic coordinates present in \code{data}.}
}
\value{

}
\description{

}
\examples{
occurrences <- read.csv(system.file("extdata", "occurrences_comp.csv",
                                    package = "ellipsenm"))

vars <- raster::stack(list.files(system.file("extdata", package = "ellipsenm"),
                                 pattern = "m_bio", full.names = TRUE))
}
