% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid_overlap.R
\name{ellipsoid_overlap}
\alias{ellipsoid_overlap}
\title{Overlap of ellipsoid-based ecological niche models}
\usage{
ellipsoid_overlap(..., overlap_type = "all", n_points = 1e+06)
}
\arguments{
\item{...}{data_overlap objects containing data for individual niches to be
compared in overlap analyses. At least two data_overlap objects are needed
to perform analyses. These objects can be created with the function
\code{\link{overlap_object}}.}

\item{overlap_type}{(character) type of overlap to be measured. Options are:
"all", "full", and "back_union". Default = "all". See details.}

\item{n_points}{}
}
\value{
A overlap_ellipsoid object containing all results from overlap analyses as
well as other information needed for plotting.
}
\description{

}
\details{
Types of overlap are as follows:

all = performs all types of overlap analyses allowed.

full = measures overlap of the complete volume of the ellipsoidal niches.

back_union = meausures overlap of ellipsoidal niches considering only the
union of the environmental conditions relevant for the two species (backgrounds).
}
\examples{
# data

# full overlap analysis

# overlap measured in the union of the two species background
}
