% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid_overlap.R
\name{ellipsoid_overlap}
\alias{ellipsoid_overlap}
\title{Overlap of ellipsoid-based ecological niche models}
\usage{
ellipsoid_overlap(..., overlap_type = "all", n_points = 1e+06)
}
\arguments{
\item{...}{data_overlap objects containing data for individual niches to be
compared in overlap analyses. At least two data_overlap objects are needed
to perform analyses. These objects can be created with the function
\code{\link{overlap_object}}.}

\item{overlap_type}{(character) type of overlap to be measured. Options are:
"all", "full", and "back_union". Default = "all". See details.}

\item{n_points}{(character) number of random points to be generated for
performing Monte-Carlo simulations (full overlap). Default = 1000000.}
}
\value{
A overlap_ellipsoid object containing all results from overlap analyses as
well as other information needed for plotting.
}
\description{
ellipsoid_overlap performs analyses to measure the degree of
overlap between two or more ellipsoid-based ecological niches as in pairwise
comparisons. Measures can be done considering the entire ellipsoid volume or
sets of environmental conditions (background).
}
\details{
Types of overlap are as follows:

all = performs all types of overlap analyses allowed.

full = measures overlap of the complete volume of the ellipsoidal niches.

back_union = meausures overlap of ellipsoidal niches considering only the
union of the environmental conditions relevant for the two species (backgrounds).
}
\examples{
# reading data
occurrences <- read.csv(system.file("extdata", "occurrences.csv",
                                    package = "ellipsenm"))

# raster layers of environmental data
vars <- raster::stack(list.files(system.file("extdata", package = "ellipsenm"),
                                 pattern = "bio", full.names = TRUE))

# preparing data
vext <- raster::extent(vars)
ext1 <- raster::extent(vext[1], (mean(vext[1:2]) + 0.2), vext[3:4])
ext2 <- raster::extent((mean(vext[1:2]) + 0.2), vext[2], vext[3:4])

# croping variables and splitting occurrences
vars1 <- raster::stack(raster::crop(vars, ext1))
vars2 <- raster::stack(raster::crop(vars, ext2))

occurrences1 <- occurrences[occurrences$longitude < (mean(vext[1:2]) + 0.2), ]
occurrences2 <- occurrences[!occurrences$longitude \%in\% occurrences1$longitude, ]

# preparing overlap objects to perform analyses
niche1 <- overlap_object(occurrences1, species =  "species", longitude = "longitude",
                         latitude = "latitude", method = "covmat", level = 95,
                         variables = vars1)

niche2 <- overlap_object(occurrences2, species =  "species", longitude = "longitude",
                         latitude = "latitude", method = "covmat", level = 95,
                         variables = vars2)

# niche overlap analysis
overlap <- ellipsoid_overlap(niche1, niche2)
}
