% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_espace.R
\name{interpretation_espace}
\alias{interpretation_espace}
\title{Helper function for interpretation of exploratory figures}
\usage{
interpretation_espace()
}
\value{
A text with help for selecting variables and methods based on exploratory plots.
}
\description{
Helper function for interpretation of exploratory figures
}
