% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overlap.R
\name{plot_overlap}
\alias{plot_overlap}
\title{Plot results of niche overlap analyses}
\usage{
plot_overlap(object, niches = c(1, 2), data = TRUE, col = c("blue",
  "red"), background = FALSE, background_type, proportion = 0.3,
  background_col = viridis::viridis, change_labels = FALSE,
  xlab = "", ylab = "", zlab = "", legend = TRUE)
}
\arguments{
\item{object}{overlap_ellipsoid object resulted from using the function
\code{\link{ellipsoid_overlap}}.}

\item{niches}{(numeric) pair of integer numbers denoting the niches to be
plotted. Default = c(1, 2).}

\item{data}{(logical) whether or not to plot points of species data. Default
= TRUE.}

\item{col}{colors to be used to plot data and ellipsoids of niches to be
compared. Default = c("blue", "red").}

\item{background}{(logical) whether or not to plot background points. Default
= TRUE.}

\item{background_type}{(character) type of background to be plotted. Options
are: "full", "back_union". See \code{\link{ellipsoid_overlap}}.}

\item{proportion}{(numeric) proportion of background to be plotted. Default = 0.3.}

\item{background_col}{color ramp to be used for coloring background points.
Default = viridis::viridis.}

\item{change_labels}{(logical) whether or not to change axes label.
Default = FALSE.}

\item{xlab}{(character) lable of x axis. Default = "".}

\item{ylab}{(character) lable of y axis. Default = "".}

\item{zlab}{(character) lable of z axis. Default = "".}

\item{legend}{(logical) whether or not to add a simple legend. Default = TRUE.}
}
\value{
A plot of the niches to be compared in environmental space.
}
\description{
plot_overlap
}
\examples{
# data

#simple plot
}
