% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Creation of HTML reports from ellipsenm main processes}
\usage{
report(report_type, name = "results_report")
}
\arguments{
\item{name}{(character) name of the HTML file to be produced.}

\item{repor_type}{(character) type of report to be produced. Options are:
"enm", "calibration", and "overlap".}
}
\value{
An HTML file summarizing results from the main processes that the ellipsenm
package performs.
}
\description{
report creates an HTML file that summarizes all outputs from
the main processes that the ellipsenm package performs.
}
\details{
This function is used along with the \code{\link{ellipsoid_model}},
ellipsoid_calibration, and ellipsoid_overlap functions.
}
