% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdata_overlap.R
\name{spdata_overlap}
\alias{spdata_overlap}
\title{Prepare species and environmental data for overlap analysis}
\usage{
spdata_overlap(method = "mve1", data, raster_layers, level = 95)
}
\arguments{
\item{method}{(character) method to construct the ellipsoid that characterizes
the species ecological niche. Available methods are: "covmat", "mve1", and
"mve2". See details of \code{\link{ellipsoid_fit}}. Default = "mve1".}

\item{data}{data.frame of occurrence records. Columns must be species,
longitude, and latitude.}

\item{raster_layers}{a RasterStack of environmental data to be used in niche
overlap ananlyses.}

\item{level}{(numeric) the confidence level of a pairwise confidence region
for the ellipsoid, expresed as percentage. Default = 95.}
}
\value{
An object of class data_overlap formated for analyses of niche overlap.
}
\description{
spdata_overlap prepares species and environmental data to perform
niche overlap analyses.
}
\examples{
# data
met <- "mve"
occurrences <- read.csv(system.file("extdata", "occurrences_thin.csv",
                                    package = "ellipsenm"))
layers <- raster::stack(list.files(system.file("extdata", package = "ellipsenm"),
                                   pattern = "m_bio", full.names = TRUE))

sp_over <- spdata_overlap(method = met, data = occurrences,
                          raster_layers = layers)
}
