% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{write_ellmeta}
\alias{write_ellmeta}
\title{Helperf function to write ellipsoid metadata}
\usage{
write_ellmeta(ellipsoid, name = "ell_meta.txt")
}
\arguments{
\item{ellipsoid}{object of class ellipsoid*.}

\item{name}{(character) name of the file to be written. If the object in
\code{ellipsoid} is replicated, number of replicate and mean, min, and max,
are added as prefixes to each file. Default = "ell_meta.txt"}
}
\value{
Writes a txt file with the ellipsoid metadata in the working directory.
}
\description{
Helperf function to write ellipsoid metadata
}
